+++ openbar-api/src/models/menu_item.rs
@@ -22,13 +22,13 @@ pub struct MenuItem {
     #[serde(rename = "picture_uri")]
     pub picture_uri: String,
     #[serde(rename = "promotion", skip_serializing_if = "Option::is_none")]
-    pub promotion: Option<i32>,
+    pub promotion: Option<i64>,
     #[serde(rename = "amount")]
-    pub amount: i32,
+    pub amount: i64,
 }
 
 impl MenuItem {
-    pub fn new(id: uuid::Uuid, name: String, picture_uri: String, amount: i32) -> MenuItem {
+    pub fn new(id: uuid::Uuid, name: String, picture_uri: String, amount: i64) -> MenuItem {
         MenuItem {
             id,
             name,
